Move autoIncrement implementation into fields package
