+++ .changeset/curly-phones-jam.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/fields': patch
+---
+
+Removed the dependency on the legacy `AutoIncrement` field implementation.
