+++ packages-next/fields/src/types/autoIncrement/index.ts
@@ -1,8 +1,7 @@
-// @ts-ignore
-import { AutoIncrement } from '@keystone-next/fields-auto-increment-legacy';
 import type { FieldType, BaseGeneratedListTypes, FieldDefaultValue } from '@keystone-next/types';
 import { resolveView } from '../../resolve-view';
 import type { FieldConfig } from '../../interfaces';
+import { AutoIncrementImplementation, PrismaAutoIncrementInterface } from './Implementation';
 
 export type AutoIncrementFieldConfig<
   TGeneratedListTypes extends BaseGeneratedListTypes
@@ -15,7 +14,11 @@ export type AutoIncrementFieldConfig<
 export const autoIncrement = <TGeneratedListTypes extends BaseGeneratedListTypes>(
   config: AutoIncrementFieldConfig<TGeneratedListTypes> = {}
 ): FieldType<TGeneratedListTypes> => ({
-  type: AutoIncrement,
+  type: {
+    type: 'AutoIncrement',
+    implementation: AutoIncrementImplementation,
+    adapters: { prisma: PrismaAutoIncrementInterface },
+  },
   config,
   views: resolveView('integer/views'),
 });
