--- packages/fields-auto-increment/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@keystone-next/fields-auto-increment-legacy",
-  "description": "KeystoneJS AutoIncrement Field Type",
-  "version": "10.0.0",
-  "author": "The KeystoneJS Development Team",
-  "license": "MIT",
-  "repository": "https://github.com/keystonejs/keystone/tree/master/packages/fields-auto-increment",
-  "homepage": "https://github.com/keystonejs/keystone",
-  "engines": {
-    "node": ">=10.0.0"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.13.10",
-    "@keystone-next/adapter-prisma-legacy": "^5.0.0",
-    "@keystone-next/fields-legacy": "^25.0.0"
-  },
-  "devDependencies": {
-    "@keystone-next/server-side-graphql-client-legacy": "*"
-  },
-  "main": "dist/fields-auto-increment-legacy.cjs.js",
-  "module": "dist/fields-auto-increment-legacy.esm.js"
-}
