--- packages/fields-auto-increment/src/index.js
@@ -1,9 +1,0 @@
-import { AutoIncrementImplementation, PrismaAutoIncrementInterface } from './Implementation';
-
-export const AutoIncrement = {
-  type: 'AutoIncrement',
-  implementation: AutoIncrementImplementation,
-  adapters: {
-    prisma: PrismaAutoIncrementInterface,
-  },
-};
